+++ examples/src/examples/index.tsx
@@ -1,25 +1,21 @@
-import KittenAndLink from  './KittenAndLink'
 import ItalicAndBold from  './ItalicAndBold'
 import BlockStyling from './BlockStyling'
 import PlainText from './PlainText'
-import Rtl from './Rtl'
 import Tree from './Tree'
 import Table from './Table'
-import InputTester from './InputTester'
 import MarkdownTest from './MarkdownTest'
 import Draggable from './Draggable'
 import TimeTravel from './TimeTravel'
 
-export {
-  PlainText,
-  Rtl,
-  KittenAndLink,
-  Table,
-  InputTester,
-  ItalicAndBold,
-  BlockStyling,
-  Tree,
-  MarkdownTest,
-  Draggable,
-  TimeTravel
+const Examples = {
+  'Plain Text': PlainText,
+  'Basic Text styling': ItalicAndBold,
+  'Block Styling': BlockStyling,
+  'Markdown': MarkdownTest,
+  'Recursive Structure': Tree,
+  'Draggable Blocks': Draggable,
+  'Table': Table,
+  'Time Travel': TimeTravel,
 }
+
+export default Examples